# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleassistance API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/zones": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Muestra todas las zonas paginadas",
                "operationId": "0f732fe44c20996cd9dd342f722cb156",
                "responses": {
                    "200": {
                        "description": "Lista de zonas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ZoneResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zones"
                ],
                "summary": "Crea una nueva zona",
                "operationId": "253822acdf31d7499579454f17b7fc8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreZoneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zona creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZoneResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zones/{id}": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Muestra una zona",
                "operationId": "9fbd7dbdda1909cd7ebc9bb4167293d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zona mostrada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZoneResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zones"
                ],
                "summary": "Actualiza una zona",
                "operationId": "28fa1be86dd47c22a90f9db32cdf9481",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateZoneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zona actualizada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZoneResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zones"
                ],
                "summary": "Elimina una zona",
                "operationId": "357ad6da21048b1eed885132d85c7c01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zona eliminada con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zones/{id}/patients": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Obtiene los pacientes de una zona",
                "operationId": "63c99002440a126a019dd9a79b74c232",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pacientes en la zona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PatientResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay pacientes asignados en esta zona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zones/{id}/teleoperators": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Obtiene los teleoperadores de una zona",
                "operationId": "5a8902f614194d61e7479dade83b77cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la zona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de teleoperadores en la zona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CallResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay teleoperadores asignados en esta zona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreZoneRequest": {
                "description": "Validación para la creación de zonas",
                "required": [
                    "city",
                    "zone"
                ],
                "properties": {
                    "city": {
                        "description": "Nombre de la ciudad",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Madrid"
                    },
                    "zone": {
                        "description": "Nombre de la zona",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centro"
                    }
                },
                "type": "object"
            },
...
            "UpdateZoneRequest": {
                "description": "Validación para la actualización de una zona",
                "required": [
                    "city",
                    "zone"
                ],
                "properties": {
                    "city": {
                        "description": "Nombre de la ciudad",
                        "type": "string",
                        "example": "Madrid"
                    },
                    "zone": {
                        "description": "Nombre de la zona dentro de la ciudad",
                        "type": "string",
                        "example": "Centro"
                    }
                },
                "type": "object"
            },
...
            "CallResource": {
                "description": "Esquema del recurso Llamada",
                "properties": {
                    "id": {
                        "description": "Identificador de la llamada",
                        "type": "integer"
                    },
                    "teleoperator": {
                        "$ref": "#/components/schemas/TeleoperatorResource"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "call_type": {
                        "description": "Tipo de llamada",
                        "type": "string",
                        "enum": [
                            "Entrante",
                            "Saliente"
                        ]
                    },
                    "type": {
                        "description": "Tipo específico de la llamada",
                        "type": "string"
                    },
                    "call_time": {
                        "description": "Fecha y hora de la llamada",
                        "type": "string",
                        "format": "date-time"
                    },
                    "title": {
                        "description": "Título de la llamada",
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descripción de la llamada",
                        "type": "string",
                        "nullable": true
                    },
                    "alert": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CallTitleResource"
                            }
                        ],
                        "nullable": true,
                        "description": "Alerta asociada a la llamada"
                    }
                },
                "type": "object"
            },
            "CallTitleResource": {
                "description": "Esquema del recurso Título de Llamada",
                "properties": {
                    "id": {
                        "description": "Identificador del título de la llamada",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Título de la llamada",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "EmergencyContactResourceRecursion": {
                "description": "Esquema del recurso Contacto de Emergencia en Recursión",
                "properties": {
                    "id": {
                        "description": "Identificador del contacto de emergencia",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del contacto de emergencia",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del contacto de emergencia",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "description": "Esquema del recurso Paciente",
                "properties": {
                    "id": {
                        "description": "Identificador del paciente",
                        "type": "integer"
                    },
                    "dni": {
                        "description": "DNI del paciente",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del paciente",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Fecha de nacimiento del paciente",
                        "type": "string",
                        "format": "date"
                    },
                    "address": {
                        "description": "Dirección del paciente",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del paciente",
                        "type": "string"
                    },
                    "health_card": {
                        "description": "Número de tarjeta sanitaria del paciente",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del paciente",
                        "type": "string",
                        "format": "email"
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "associated_teleoperator_id": {
                        "$ref": "#/components/schemas/TeleoperatorResource"
                    },
                    "personal_situation": {
                        "description": "Situación personal del paciente",
                        "type": "string"
                    },
                    "health_situation": {
                        "description": "Situación de salud del paciente",
                        "type": "string"
                    },
                    "housing_situation": {
                        "description": "Situación de vivienda del paciente",
                        "type": "string"
                    },
                    "economic_situation": {
                        "description": "Situación económica del paciente",
                        "type": "string"
                    },
                    "autonomy": {
                        "description": "Nivel de autonomía del paciente",
                        "type": "string"
                    },
                    "emergency_contacts": {
                        "description": "Contactos de emergencia asociados al paciente",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmergencyContactResourceRecursion"
                        }
                    }
                },
                "type": "object"
            },
...
            "TeleoperatorResource": {
                "description": "Esquema del recurso Teleoperador",
                "properties": {
                    "id": {
                        "description": "Identificador del teleoperador",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del teleoperador",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del teleoperador",
                        "type": "string",
                        "format": "email"
                    },
                    "prefix": {
                        "description": "Prefijo telefónico",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Número de teléfono del teleoperador",
                        "type": "string"
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "languages": {
                        "description": "Idiomas hablados por el teleoperador",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "hiring_date": {
                        "description": "Fecha de contratación",
                        "type": "string",
                        "format": "date"
                    },
                    "code": {
                        "description": "Código de identificación del teleoperador",
                        "type": "string"
                    },
                    "firing_date": {
                        "description": "Fecha de despido (puede ser nula si sigue activo)",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ZoneResource": {
                "description": "Esquema del recurso Zona",
                "properties": {
                    "id": {
                        "description": "Identificador de la zona",
                        "type": "integer"
                    },
                    "city": {
                        "description": "Nombre de la ciudad",
                        "type": "string"
                    },
                    "zone": {
                        "description": "Nombre de la zona dentro de la ciudad",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}